+++ AppCenter/AppCenterTests/MSACAppCenterTests.m
@@ -1017,4 +1017,70 @@ static NSString *const kMSACNullifiedInstallIdString = @"00000000-0000-0000-0000
   XCTAssertNil([[MSACUserIdContext sharedInstance] userIdAt:[NSDate dateWithTimeIntervalSince1970:5000]]);
 }
 
+- (void)testDisableNetworkRequestsBeforeStart {
+
+  // If
+  NSArray *services = @[ MSACMockService.class, MSACMockSecondService.class ];
+
+  // When
+  [self.settingsMock setObject:@NO forKey:kMSACAppCenterNetworkRequestsAllowedKey];
+  [MSACAppCenter start:MSAC_UUID_STRING withServices:services];
+  OCMReject([self.channelGroupMock pauseWithIdentifyingObject:OCMOCK_ANY]);
+  OCMReject([self.channelGroupMock resumeWithIdentifyingObject:OCMOCK_ANY]);
+
+  // Then
+  XCTAssertFalse([MSACAppCenter isNetworkRequestsAllowed]);
+  XCTAssertNotNil([[MSACAppCenter sharedInstance] appSecret]);
+  XCTAssertTrue([MSACMockService sharedInstance].started);
+  XCTAssertTrue([MSACMockSecondService sharedInstance].started);
+  OCMVerifyAll(self.channelGroupMock);
+}
+
+- (void)testDisableNetworkRequestsAfterStart {
+
+  // If
+  __block NSInteger pauseInvoked = 0;
+  __block NSInteger resumeInvoked = 0;
+  __block NSObject *tokenObject = nil;
+  NSArray *services = @[ MSACMockService.class, MSACMockSecondService.class ];
+  OCMStub([self.channelGroupMock pauseWithIdentifyingObject:OCMOCK_ANY]).andDo(^(NSInvocation *invocation) {
+    [invocation getArgument:&tokenObject atIndex:0];
+    pauseInvoked++;
+  });
+  OCMStub([self.channelGroupMock resumeWithIdentifyingObject:OCMOCK_ANY]).andDo(^(NSInvocation *invocation) {
+    [invocation getArgument:&tokenObject atIndex:0];
+    resumeInvoked++;
+  });
+
+  // When
+  [MSACAppCenter start:MSAC_UUID_STRING withServices:services];
+
+  // Then
+  XCTAssertTrue([MSACAppCenter isNetworkRequestsAllowed]);
+  XCTAssertNotNil([[MSACAppCenter sharedInstance] appSecret]);
+  XCTAssertTrue([MSACMockService sharedInstance].started);
+  XCTAssertTrue([MSACMockSecondService sharedInstance].started);
+  XCTAssertEqual(pauseInvoked, 0);
+  XCTAssertEqual(resumeInvoked, 0);
+  XCTAssertNil(tokenObject);
+
+  // When
+  [MSACAppCenter setNetworkRequestsAllowed:NO];
+
+  // Then
+  XCTAssertFalse([MSACAppCenter isNetworkRequestsAllowed]);
+  XCTAssertEqual(pauseInvoked, 1);
+  XCTAssertEqual(resumeInvoked, 0);
+  XCTAssertEqual(tokenObject, self.channelGroupMock);
+
+  // When
+  [MSACAppCenter setNetworkRequestsAllowed:YES];
+
+  // Then
+  XCTAssertTrue([MSACAppCenter isNetworkRequestsAllowed]);
+  XCTAssertEqual(pauseInvoked, 1);
+  XCTAssertEqual(resumeInvoked, 1);
+  XCTAssertEqual(tokenObject, self.channelGroupMock);
+}
+
 @end